Implement exclusive selection of event intervals
The intersection of event intervals and user selection is made on every
update for a seamless selection experience.

Due to the differing heights of each selection due to split event
intervals, the global selection duration had to be removed.

Instead, a fade effect is used to add some continuity to the selections.