Standardise API call structure
Due to the nature of SSR with Sapper, the `fetch` context might not
always be the same.

Sapper provides a context-aware `fetch` function that will forward
browser credentials if necessary during hydration. Therefore, all API
calls should have a dynamic `fetch` context.

In addition, credentials for creating events and adding users to events
are included to allow automatic login.