Refactor how schedules are displayed
By thinking of schedules as uninterupted intervals of time, we can
remove the need for splitting intervals across the midnight line.
Instead, visualize the midnight line as a horizontal divide across the
columns, but allow selections to be made across the midnight boundary.

This way, regardless of timezones, the entire interval will be
contiguous.